improve binary building
