+++ internal/cli/cmd/mcp_windows_test.go
@@ -1,0 +1,32 @@
+// Copyright (c) 2026 Lateralus Labs, LLC.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build windows
+
+package cmd
+
+import (
+	"os/exec"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestSetSysProcAttr(t *testing.T) {
+	cmd := exec.Command("echo", "test")
+	setSysProcAttr(cmd)
+
+	// On Windows, setSysProcAttr sets CreationFlags for process group management
+	assert.NotNil(t, cmd.SysProcAttr)
+	assert.NotZero(t, cmd.SysProcAttr.CreationFlags)
+}
